# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/users": {
      "post": {
        "operationId": "UsersController_create",
        "summary": "Create a new user",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The user has been successfully created."
          },
          "400": {
            "description": "Bad Request."
          }
        },
        "tags": [
          "users"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "UsersController_findAll",
        "summary": "Get all users",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Return all users."
          },
          "404": {
            "description": "Users not found."
          }
        },
        "tags": [
          "users"
        ]
      }
    },
    "/users/{id}": {
      "get": {
        "operationId": "UsersController_findOne",
        "summary": "Get a user by id",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return a user."
          },
          "404": {
            "description": "User not found."
          }
        },
        "tags": [
          "users"
        ]
      },
      "put": {
        "operationId": "UsersController_update",
        "summary": "Update a user by id",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The user has been successfully updated."
          },
          "404": {
            "description": "User not found."
          }
        },
        "tags": [
          "users"
        ]
      },
      "delete": {
        "operationId": "UsersController_remove",
        "summary": "Delete a user by id",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The user has been successfully deleted."
          },
          "404": {
            "description": "User not found."
          }
        },
        "tags": [
          "users"
        ]
      }
    },
...
    "/tasks": {
      "post": {
        "operationId": "TasksController_create",
        "summary": "Create a new task",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTaskDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The task has been successfully created."
          },
          "400": {
            "description": "Bad Request."
          }
        },
        "tags": [
          "tasks"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "TasksController_findAll",
        "summary": "Get all tasks",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Return all tasks."
          },
          "404": {
            "description": "Tasks not found."
          }
        },
        "tags": [
          "tasks"
        ]
      }
    },
    "/tasks/{id}": {
      "get": {
        "operationId": "TasksController_findOne",
        "summary": "Get a task by id",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return a task."
          },
          "404": {
            "description": "Task not found."
          }
        },
        "tags": [
          "tasks"
        ]
      },
      "put": {
        "operationId": "TasksController_update",
        "summary": "Update a task by id",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTaskDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The task has been successfully updated."
          },
          "404": {
            "description": "Task not found."
          }
        },
        "tags": [
          "tasks"
        ]
      },
      "delete": {
        "operationId": "TasksController_remove",
        "summary": "Delete a task by id",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The task has been successfully deleted."
          },
          "404": {
            "description": "Task not found."
          }
        },
        "tags": [
          "tasks"
        ]
      }
    },
    "/tasks-categories": {
      "post": {
        "operationId": "TaskCategoriesController_create",
        "summary": "Create a new task category",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTaskCategorieDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The task category has been successfully created."
          },
          "400": {
            "description": "Bad Request."
          }
        },
        "tags": [
          "tasks-categories"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "TaskCategoriesController_findAll",
        "summary": "Get all task categories",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Return all task categories."
          },
          "404": {
            "description": "Task categories not found."
          }
        },
        "tags": [
          "tasks-categories"
        ]
      }
    },
    "/tasks-categories/{id}": {
      "get": {
        "operationId": "TaskCategoriesController_findOne",
        "summary": "Get a task category by id",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return a task category."
          },
          "404": {
            "description": "Task category not found."
          }
        },
        "tags": [
          "tasks-categories"
        ]
      },
      "put": {
        "operationId": "TaskCategoriesController_update",
        "summary": "Update a task category by id",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTaskCategorieDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The task category has been successfully updated."
          },
          "404": {
            "description": "Task category not found."
          }
        },
        "tags": [
          "tasks-categories"
        ]
      },
      "delete": {
        "operationId": "TaskCategoriesController_remove",
        "summary": "Delete a task category by id",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The task category has been successfully deleted."
          },
          "404": {
            "description": "Task category not found."
          }
        },
        "tags": [
          "tasks-categories"
        ]
      }
    },
    "/tasks-categories/{id}/tasks": {
      "get": {
        "operationId": "TaskCategoriesController_findTasks",
        "summary": "Get all tasks for a specific task category",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return all tasks for the specified task category."
          },
          "404": {
            "description": "Tasks not found."
          }
        },
        "tags": [
          "tasks-categories"
        ]
      }
    },
...
  },
  "info": {
    "title": "Api ZenFlow",
    "description": "Api for a CRM/Forum application with modules.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "description": "The status of the user",
            "enum": [
              "Verified",
              "NotVerified",
              "Suspended",
              "Banned"
            ]
          },
          "role": {
            "type": "string",
            "description": "The role of the user",
            "enum": [
              "USER",
              "ADMIN",
              "MEGA_ADMIN"
            ]
          },
          "token": {
            "type": "string",
            "description": "The token of the user"
          },
          "uid": {
            "type": "string",
            "description": "The UID of the user"
          }
        }
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "description": "The status of the user",
            "enum": [
              "Verified",
              "NotVerified",
              "Suspended",
              "Banned"
            ]
          },
          "role": {
            "type": "string",
            "description": "The role of the user",
            "enum": [
              "USER",
              "ADMIN",
              "MEGA_ADMIN"
            ]
          },
          "token": {
            "type": "string",
            "description": "The token of the user"
          },
          "uid": {
            "type": "string",
            "description": "The UID of the user"
          }
        }
      },
...
      "CreateTaskDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "The title of the task"
          },
          "description": {
            "type": "string",
            "description": "The description of the task"
          },
          "taskCategoryId": {
            "type": "string",
            "description": "The ID of the task category"
          },
          "assigneeId": {
            "type": "string",
            "description": "The ID of the assignee"
          },
          "status": {
            "type": "string",
            "description": "The status of the task"
          },
          "completed": {
            "type": "boolean",
            "description": "Whether the task is completed"
          },
          "dueDate": {
            "format": "date-time",
            "type": "string",
            "description": "The due date of the task"
          },
          "priority": {
            "type": "string",
            "description": "The priority of the task"
          },
          "comments": {
            "description": "Comments on the task",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "flags": {
            "description": "Flags associated with the task",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "timeEntries": {
            "description": "Time entries for the task",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "customFields": {
            "type": "object",
            "description": "Custom fields for the task"
          },
          "parentTaskId": {
            "type": "string",
            "description": "The ID of the parent task"
          },
          "subTasks": {
            "description": "Sub-tasks associated with the task",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "section": {
            "type": "string",
            "description": "The section of the task"
          },
          "order": {
            "type": "number",
            "description": "The order of the task"
          }
        },
        "required": [
          "title",
          "description",
          "status"
        ]
      },
      "UpdateTaskDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "The title of the task"
          },
          "description": {
            "type": "string",
            "description": "The description of the task"
          },
          "taskCategoryId": {
            "type": "string",
            "description": "The ID of the task category"
          },
          "assigneeId": {
            "type": "string",
            "description": "The ID of the assignee"
          },
          "status": {
            "type": "string",
            "description": "The status of the task"
          },
          "completed": {
            "type": "boolean",
            "description": "Whether the task is completed"
          },
          "dueDate": {
            "format": "date-time",
            "type": "string",
            "description": "The due date of the task"
          },
          "priority": {
            "type": "string",
            "description": "The priority of the task"
          },
          "comments": {
            "description": "Comments on the task",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "flags": {
            "description": "Flags associated with the task",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "timeEntries": {
            "description": "Time entries for the task",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "customFields": {
            "type": "object",
            "description": "Custom fields for the task"
          },
          "parentTaskId": {
            "type": "string",
            "description": "The ID of the parent task"
          },
          "subTasks": {
            "description": "Sub-tasks associated with the task",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "section": {
            "type": "string",
            "description": "The section of the task"
          },
          "order": {
            "type": "number",
            "description": "The order of the task"
          }
        }
      },
      "CreateTaskCategorieDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "The name of the task category"
          },
          "description": {
            "type": "string",
            "description": "The description of the task category"
          }
        },
        "required": [
          "name",
          "description"
        ]
      },
      "UpdateTaskCategorieDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "The name of the task category"
          },
          "description": {
            "type": "string",
            "description": "The description of the task category"
          }
        }
      },
...
    }
  }
}